Neon/requirements.txt
discord.py>=2.3.0
lyricsgenius
flask
python-dotenv
